[CORDA-1799]: Moving towards resource files in TestCorDappBuilder.
